- Checksum: contient le code nécessaire pour calculer un checksum "fort" (MD5). - Ascii85: les leaders/trailers sont facultatifs <~...~>. - ImageDataStore: code fonctionnel; on peut stocker une image dans la base de données, puis relire les données en retour.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010/Epsitec.Cresus@14844 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.Support/FolderItemIconCache.cs
@@ -31,7 +31,7 @@ namespace Epsitec.Common.Support
 			//	calling Release, so be very cautious in this method !
 			
 			byte[] sourceData = image.BitmapImage.GetRawBitmapBytes ();
-			long   sourceCrc  = IO.Checksum.ComputeCrc32 (delegate (IO.IChecksum checksum) { checksum.Update (sourceData); });
+			int    sourceCrc  = IO.Checksum.ComputeCrc32 (delegate (IO.IChecksum checksum) { checksum.Update (sourceData); });
 
 			Item[] items;
 			Item matchingItem = new Item ();
@@ -157,7 +157,7 @@ namespace Epsitec.Common.Support
 		private struct Item
 		{
 			public Drawing.Image image;
-			public long crc;
+			public int crc;
 			public long id;
 			public int count;
 		}
